refactor: remove redundant file, switch to new framework

+++ src/observer/sql/table/schema.cpp
@@ -2,7 +2,7 @@
 
 Schema::Schema(const std::vector<Column> &columns)
 {
-  uint32_t curr_offset = 0;
+  size_t curr_offset = 0;
   for (auto column : columns) {
     // set column offset
     column.column_offset_ = curr_offset;
